chore(core): Split source control service (#18784)
