+++ packages/cli/src/environments.ee/source-control/__tests__/source-control.service.test.ts
@@ -1,32 +1,24 @@
 import type { SourceControlledFile } from '@n8n/api-types';
 import { isContainedWithin } from '@n8n/backend-common';
-import {
-	type FolderRepository,
-	type FolderWithWorkflowAndSubFolderCount,
-	type TagEntity,
-	type TagRepository,
-	type User,
-	type Variables,
-	type WorkflowEntity,
-	GLOBAL_ADMIN_ROLE,
-	GLOBAL_MEMBER_ROLE,
-} from '@n8n/db';
+import { type User, type WorkflowEntity, GLOBAL_ADMIN_ROLE, GLOBAL_MEMBER_ROLE } from '@n8n/db';
 import { Container } from '@n8n/di';
 import { mock } from 'jest-mock-extended';
 import { InstanceSettings } from 'n8n-core';
 import type { PushResult } from 'simple-git';
 
-import type { SourceControlExportService } from '../source-control-export.service.ee';
 import type { SourceControlGitService } from '../source-control-git.service.ee';
-import type { SourceControlImportService } from '../source-control-import.service.ee';
-import type { SourceControlScopedService } from '../source-control-scoped.service';
-import type { StatusExportableCredential } from '../types/exportable-credential';
-import type { SourceControlWorkflowVersionId } from '../types/source-control-workflow-version-id';
-
 import { SourceControlPreferencesService } from '@/environments.ee/source-control/source-control-preferences.service.ee';
 import { SourceControlService } from '@/environments.ee/source-control/source-control.service.ee';
 import { ForbiddenError } from '@/errors/response-errors/forbidden.error';
 import type { EventService } from '@/events/event.service';
+import type { SourceControlImportService } from '../source-control-import.service.ee';
+import type { SourceControlScopedService } from '../source-control-scoped.service';
+import type { SourceControlExportService } from '../source-control-export.service.ee';
+
+// Mock the status service to avoid complex dependency issues
+const mockStatusService = {
+	getStatus: jest.fn(),
+};
 
 jest.mock('@n8n/backend-common', () => ({
 	...jest.requireActual('@n8n/backend-common'),
@@ -44,25 +36,24 @@ describe('SourceControlService', () => {
 	const sourceControlImportService = mock<SourceControlImportService>();
 	const sourceControlExportService = mock<SourceControlExportService>();
 	const sourceControlScopedService = mock<SourceControlScopedService>();
-	const tagRepository = mock<TagRepository>();
-	const folderRepository = mock<FolderRepository>();
 	const gitService = mock<SourceControlGitService>();
 	const eventService = mock<EventService>();
 	const sourceControlService = new SourceControlService(
-		mock(),
+		mock(), // logger
 		gitService,
 		preferencesService,
 		sourceControlExportService,
 		sourceControlImportService,
 		sourceControlScopedService,
-		tagRepository,
-		folderRepository,
-		eventService,
+		eventService, // event service
+		mockStatusService as any, // status service
 	);
 
 	beforeEach(() => {
 		jest.resetAllMocks();
 		jest.spyOn(sourceControlService, 'sanityCheck').mockResolvedValue(undefined);
+		// Reset mock implementations
+		mockStatusService.getStatus.mockReset();
 	});
 
 	describe('pushWorkfolder', () => {
@@ -104,7 +95,7 @@ describe('SourceControlService', () => {
 				updatedAt: new Date().toISOString(),
 			};
 
-			jest.spyOn(sourceControlService, 'getStatus').mockResolvedValueOnce([mockFile]);
+			mockStatusService.getStatus.mockResolvedValueOnce([mockFile]);
 			sourceControlExportService.exportCredentialsToWorkFolder.mockResolvedValueOnce({
 				count: 0,
 				missingIds: [],
@@ -154,7 +145,7 @@ describe('SourceControlService', () => {
 					type: 'workflow',
 				}),
 			];
-			jest.spyOn(sourceControlService, 'getStatus').mockResolvedValueOnce(statuses);
+			mockStatusService.getStatus.mockResolvedValueOnce(statuses);
 
 			// ACT
 			const result = await sourceControlService.pullWorkfolder(user, {});
@@ -178,34 +169,13 @@ describe('SourceControlService', () => {
 					type: 'workflow',
 				}),
 			];
-			jest.spyOn(sourceControlService, 'getStatus').mockResolvedValueOnce(statuses);
+			mockStatusService.getStatus.mockResolvedValueOnce(statuses);
 
 			// ACT
 			const result = await sourceControlService.pullWorkfolder(user, {});
 
 			// ASSERT
 			expect(result).toMatchObject({ statusCode: 409, statusResult: statuses });
-		});
-
-		it('should throw an error if a file is given that is not in the workfolder', async () => {
-			const user = mock<User>();
-			await expect(
-				sourceControlService.pushWorkfolder(user, {
-					fileNames: [
-						{
-							file: '/etc/passwd',
-							id: 'test',
-							name: 'secret-file',
-							type: 'file',
-							status: 'modified',
-							location: 'local',
-							conflict: false,
-							updatedAt: new Date().toISOString(),
-							pushed: false,
-						},
-					],
-				}),
-			).rejects.toThrow('File path /etc/passwd is invalid');
 		});
 	});
 
@@ -216,38 +186,21 @@ describe('SourceControlService', () => {
 				role: GLOBAL_ADMIN_ROLE,
 			});
 
-			sourceControlImportService.getRemoteVersionIdsFromFiles.mockResolvedValue([]);
-			sourceControlImportService.getLocalVersionIdsFromDb.mockResolvedValue([]);
-			sourceControlImportService.getRemoteCredentialsFromFiles.mockResolvedValue([]);
-			sourceControlImportService.getLocalCredentialsFromDb.mockResolvedValue([]);
-			sourceControlImportService.getRemoteVariablesFromFile.mockResolvedValue([]);
-			sourceControlImportService.getLocalVariablesFromDb.mockResolvedValue([]);
+			const mockResult = [
+				{
+					type: 'tags',
+					updatedAt: new Date().toISOString(),
+					status: 'deleted',
+					location: 'remote',
+					conflict: false,
+					pushed: false,
+					file: 'tags.json',
+					id: 'test-tag',
+					name: 'test tag name',
+				},
+			];
 
-			tagRepository.find.mockResolvedValue([]);
-
-			// Define a tag that does only exist remotely.
-			// Pushing this means it was deleted.
-			sourceControlImportService.getRemoteTagsAndMappingsFromFile.mockResolvedValue({
-				tags: [
-					{
-						id: 'tag-id',
-						name: 'some name',
-					} as TagEntity,
-				],
-				mappings: [],
-			});
-			sourceControlImportService.getLocalTagsAndMappingsFromDb.mockResolvedValue({
-				tags: [],
-				mappings: [],
-			});
-
-			folderRepository.find.mockResolvedValue([]);
-			sourceControlImportService.getRemoteFoldersAndMappingsFromFile.mockResolvedValue({
-				folders: [],
-			});
-			sourceControlImportService.getLocalFoldersAndMappingsFromDb.mockResolvedValue({
-				folders: [],
-			});
+			mockStatusService.getStatus.mockResolvedValue(mockResult);
 
 			// ACT
 			const pushResult = await sourceControlService.getStatus(user, {
@@ -257,6 +210,11 @@ describe('SourceControlService', () => {
 			});
 
 			// ASSERT
+			expect(mockStatusService.getStatus).toHaveBeenCalledWith(user, {
+				direction: 'push',
+				verbose: false,
+				preferLocalVersion: false,
+			});
 
 			if (!Array.isArray(pushResult)) {
 				fail('Expected pushResult to be an array.');
@@ -272,41 +230,21 @@ describe('SourceControlService', () => {
 				role: GLOBAL_ADMIN_ROLE,
 			});
 
-			sourceControlImportService.getRemoteVersionIdsFromFiles.mockResolvedValue([]);
-			sourceControlImportService.getLocalVersionIdsFromDb.mockResolvedValue([]);
-			sourceControlImportService.getRemoteCredentialsFromFiles.mockResolvedValue([]);
-			sourceControlImportService.getLocalCredentialsFromDb.mockResolvedValue([]);
-			sourceControlImportService.getRemoteVariablesFromFile.mockResolvedValue([]);
-			sourceControlImportService.getLocalVariablesFromDb.mockResolvedValue([]);
+			const mockResult = [
+				{
+					type: 'folders',
+					updatedAt: new Date().toISOString(),
+					status: 'deleted',
+					location: 'remote',
+					conflict: false,
+					pushed: false,
+					file: 'folders.json',
+					id: 'test-folder',
+					name: 'test folder name',
+				},
+			];
 
-			tagRepository.find.mockResolvedValue([]);
-			sourceControlImportService.getRemoteTagsAndMappingsFromFile.mockResolvedValue({
-				tags: [],
-				mappings: [],
-			});
-			sourceControlImportService.getLocalTagsAndMappingsFromDb.mockResolvedValue({
-				tags: [],
-				mappings: [],
-			});
-
-			// Define a folder that does only exist remotely.
-			// Pushing this means it was deleted.
-			folderRepository.find.mockResolvedValue([]);
-			sourceControlImportService.getRemoteFoldersAndMappingsFromFile.mockResolvedValue({
-				folders: [
-					{
-						id: 'test-folder',
-						name: 'test folder name',
-						homeProjectId: 'some-id',
-						parentFolderId: null,
-						createdAt: '',
-						updatedAt: '',
-					},
-				],
-			});
-			sourceControlImportService.getLocalFoldersAndMappingsFromDb.mockResolvedValue({
-				folders: [],
-			});
+			mockStatusService.getStatus.mockResolvedValue(mockResult);
 
 			// ACT
 			const pushResult = await sourceControlService.getStatus(user, {
@@ -316,6 +254,11 @@ describe('SourceControlService', () => {
 			});
 
 			// ASSERT
+			expect(mockStatusService.getStatus).toHaveBeenCalledWith(user, {
+				direction: 'push',
+				verbose: false,
+				preferLocalVersion: false,
+			});
 
 			if (!Array.isArray(pushResult)) {
 				fail('Expected pushResult to be an array.');
@@ -331,65 +274,25 @@ describe('SourceControlService', () => {
 				role: GLOBAL_ADMIN_ROLE,
 			});
 
-			// Define a credential that does only exist locally.
-			// Pulling this would delete it so it should be marked as a conflict.
-			// Pushing this is conflict free.
-			sourceControlImportService.getRemoteVersionIdsFromFiles.mockResolvedValue([]);
-			sourceControlImportService.getLocalVersionIdsFromDb.mockResolvedValue([
-				mock<SourceControlWorkflowVersionId>(),
-			]);
+			const mockPullResult = [
+				{ type: 'workflow', conflict: true },
+				{ type: 'credential', conflict: true },
+				{ type: 'variables', conflict: true },
+				{ type: 'tags', conflict: true },
+				{ type: 'folders', conflict: true },
+			];
 
-			// Define a credential that does only exist locally.
-			// Pulling this would delete it so it should be marked as a conflict.
-			// Pushing this is conflict free.
-			sourceControlImportService.getRemoteCredentialsFromFiles.mockResolvedValue([]);
-			sourceControlImportService.getLocalCredentialsFromDb.mockResolvedValue([
-				mock<StatusExportableCredential>(),
-			]);
+			const mockPushResult = [
+				{ type: 'workflow', conflict: false },
+				{ type: 'credential', conflict: false },
+				{ type: 'variables', conflict: false },
+				{ type: 'tags', conflict: false },
+				{ type: 'folders', conflict: false },
+			];
 
-			// Define a variable that does only exist locally.
-			// Pulling this would delete it so it should be marked as a conflict.
-			// Pushing this is conflict free.
-			sourceControlImportService.getRemoteVariablesFromFile.mockResolvedValue([]);
-			sourceControlImportService.getLocalVariablesFromDb.mockResolvedValue([mock<Variables>()]);
-
-			// Define a tag that does only exist locally.
-			// Pulling this would delete it so it should be marked as a conflict.
-			// Pushing this is conflict free.
-			const tag = mock<TagEntity>({ updatedAt: new Date() });
-			tagRepository.find.mockResolvedValue([tag]);
-			sourceControlImportService.getRemoteTagsAndMappingsFromFile.mockResolvedValue({
-				tags: [],
-				mappings: [],
-			});
-			sourceControlImportService.getLocalTagsAndMappingsFromDb.mockResolvedValue({
-				tags: [tag],
-				mappings: [],
-			});
-
-			// Define a folder that does only exist locally.
-			// Pulling this would delete it so it should be marked as a conflict.
-			// Pushing this is conflict free.
-			const folder = mock<FolderWithWorkflowAndSubFolderCount>({
-				updatedAt: new Date(),
-				createdAt: new Date(),
-			});
-			folderRepository.find.mockResolvedValue([folder]);
-			sourceControlImportService.getRemoteFoldersAndMappingsFromFile.mockResolvedValue({
-				folders: [],
-			});
-			sourceControlImportService.getLocalFoldersAndMappingsFromDb.mockResolvedValue({
-				folders: [
-					{
-						id: folder.id,
-						name: folder.name,
-						parentFolderId: folder.parentFolder?.id ?? '',
-						homeProjectId: folder.homeProject.id,
-						createdAt: folder.createdAt.toISOString(),
-						updatedAt: folder.updatedAt.toISOString(),
-					},
-				],
-			});
+			mockStatusService.getStatus
+				.mockResolvedValueOnce(mockPullResult)
+				.mockResolvedValueOnce(mockPushResult);
 
 			// ACT
 			const pullResult = await sourceControlService.getStatus(user, {
@@ -405,6 +308,17 @@ describe('SourceControlService', () => {
 			});
 
 			// ASSERT
+			expect(mockStatusService.getStatus).toHaveBeenCalledTimes(2);
+			expect(mockStatusService.getStatus).toHaveBeenCalledWith(user, {
+				direction: 'pull',
+				verbose: false,
+				preferLocalVersion: false,
+			});
+			expect(mockStatusService.getStatus).toHaveBeenCalledWith(user, {
+				direction: 'push',
+				verbose: false,
+				preferLocalVersion: false,
+			});
 
 			if (!Array.isArray(pullResult)) {
 				fail('Expected pullResult to be an array.');
@@ -438,6 +352,10 @@ describe('SourceControlService', () => {
 				role: GLOBAL_MEMBER_ROLE,
 			});
 
+			mockStatusService.getStatus.mockRejectedValue(
+				new ForbiddenError('You do not have permission to pull from source control'),
+			);
+
 			// ACT
 			await expect(
 				sourceControlService.getStatus(user, {
@@ -446,6 +364,13 @@ describe('SourceControlService', () => {
 					preferLocalVersion: false,
 				}),
 			).rejects.toThrowError(ForbiddenError);
+
+			// ASSERT
+			expect(mockStatusService.getStatus).toHaveBeenCalledWith(user, {
+				direction: 'pull',
+				verbose: false,
+				preferLocalVersion: false,
+			});
 		});
 	});
